+++ DSCResources/MSFT_CertReq/MSFT_CertReq.psm1
@@ -811,13 +811,13 @@ function Test-TargetResource
             $($LocalizedData.TestingCertReqStatusMessage -f $Subject, $ca)
         ) -join '' )
 
-    # Exception for standard template DomainControllerAuthentication
     $cert = Get-Childitem -Path Cert:\LocalMachine\My |
         Where-Object -FilterScript {
-        $_.Subject -eq $Subject -and `
+            (Compare-CertificateSubject -ReferenceSubject $_.Subject -DifferenceSubject $Subject) -and `
             $_.Issuer.split(',')[0] -eq "CN=$CARootName"
     }
 
+    # Exception for standard template DomainControllerAuthentication
     if ($CertificateTemplate -eq 'DomainControllerAuthentication')
     {
         $cert = Get-Childitem -Path Cert:\LocalMachine\My |
@@ -940,3 +940,40 @@ function Test-TargetResource
         ) -join '' )
     return $false
 } # end function Test-TargetResource
+
+<#
+    .SYNOPSIS
+    Compares two certificate subjects.
+
+    .PARAMETER ReferenceSubject
+    The certificate subject to compare.
+
+    .PARAMETER DifferenceSubject
+    The certificate subject to compare with the ReferenceSubject.
+#>
+function Compare-CertificateSubject
+{
+    [CmdletBinding()]
+    [OutputType([System.Boolean])]
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
+        [System.String]
+        $ReferenceSubject,
+
+        [Parameter(Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
+        [System.String]
+        $DifferenceSubject
+    )
+
+    $referenceSubjectArray = ($ReferenceSubject -split ',').Trim() | Sort-Object
+    $differenceSubjectArray = ($DifferenceSubject -split ',').Trim() | Sort-Object
+
+    $difference = Compare-Object `
+        -ReferenceObject $referenceSubjectArray `
+        -DifferenceObject $differenceSubjectArray
+
+    return ($difference.Count -eq 0)
+}